Fix service process handling
The old combination of Type, KillMode and PIDFile makes the service to
leave behind some running python child processes, as the systemd unit
file is explicitly configured to only kill the launching shell script.

By deleting the KillMode (leaving it to the default of control-group, so
systemd kills all the processes instead only the parent) and PIDFile (as
we just don't care about it since systemd knows how to handle foreground
processes with children) options, the service behaves as it should.